test(compiler-plugin/previewCollection): move logic-spanning e2e tests under previewCollection/
Ticket 1 で確定した production の `feature/previewCollection/` 構造に sync する。
本 commit では複数 logic を横断する e2e テスト群 (collect の挙動全般 / cross-module
aggregation / Kotlin version gate / scope routing / enabled flag) を
`feature/previewCollection/` 直下に再配置する。logic 別 (fir/hintGeneration や
ir/collectPreviewsReplacement) に閉じきらない動作確認なので feature 直下を据える。

移動 (9 files):

- `AggregationTest.kt` (collect e2e)
- `CollectPreviewsTest.kt` (collect e2e)
- `CollectedPreviewMetadataTest.kt` (collect e2e)
- `CrossModuleAggregationTest.kt` (cross-module e2e)
- `CrossModuleAggregationKlibTest.kt` (cross-module e2e / klib)
- `PreviewDiscoveryTest.kt` (@Preview detection e2e)
- `PreviewHintEnabledFlagTest.kt` (enabled flag emit/collect e2e; PR #186)
- `PreviewHintScopeTest.kt` (scope-aware emit + IR collect e2e; PR #187)
- `UnsupportedKotlinErrorTest.kt` (Kotlin version gate e2e)

package: `feature` → `feature.previewCollection`

`PreviewHintEnabledFlagTest` / `PreviewHintScopeTest` は同 source set 内の
`PreviewHintTestSupport` (`internal` extension functions on `File`) に依存する
ため `import me.tbsten.compose.preview.lab.compiler.feature.previewHint*`
を追加。PreviewHintTestSupport 自体は別 commit で `utils/` 配下へ移動予定。

検証:
- `:compiler-plugin:ktlintFormat` + `:compiler-plugin:ktlintCheck` pass
- `:compiler-plugin:test --rerun-tasks` pass (131 tests total)
- `apiCheck` pass
- `find compiler-plugin -name build/api -type d` で `compat-k*/build/api/` 以外ゼロ
- wildcard import 増加なし
